requirements.txt
django==1.5.3
pymongo==2.5
mongoengine==0.8.2
blinker==1.2
python-dateutil==1.5
pytz==2012j
reportlab==2.6
django-celery==3.0.17
celery-with-redis==3.0
django-tastypie==0.9.15
django-gravatar2==1.0.4
django-storages==1.1.8
boto==2.13.0
sphinx==1.1.3
-e git+git://github.com/Naeka/django-tastypie-mongoengine.git@closer-queryset-api#egg=tastypie_mongoengine
-e git+git://github.com/Naeka/pyes.git#egg=pyes
six  # for pyes
gunicorn==0.17.2
watchdog==0.6.0
dj-database-url==0.2.1
-e git+git://github.com/Naeka/icalendar.git#egg=icalendar
vobject==0.8.1c
MySQL-python==1.2.4
newrelic==2.0.0.1
pylibmc==1.2.3
django-pylibmc-sasl==0.2.4
lxml==3.0.2
defusedxml==0.4.1
PyYAML==3.10
biplist==0.5
South==0.7.6
inflection==0.1.2
pygments==1.6
pusher==0.7
johnny-cache==1.4
Pillow==2.1.0
django-cors-headers==0.11
raven
